Fix message handling logic + move NewBot into config.go
