+++ access/slack/bot.go
@@ -19,7 +19,6 @@ package main
 import (
 	"context"
 	"encoding/json"
-	"net/http"
 	"net/url"
 	"time"
 
@@ -43,48 +42,6 @@ type SlackBot struct {
 	client      *resty.Client
 	clusterName string
 	webProxyURL *url.URL
-}
-
-// NewBot initializes the new Slack message generator (SlackBot)
-// takes GenericAPIConfig as an argument.
-func (c *SlackConfig) NewBot(clusterName, webProxyAddr string) (common.MessagingBot, error) {
-	var (
-		webProxyURL *url.URL
-		err         error
-	)
-	if webProxyAddr != "" {
-		if webProxyURL, err = lib.AddrToURL(webProxyAddr); err != nil {
-			return SlackBot{}, trace.Wrap(err)
-		}
-	}
-
-	token := "Bearer " + c.Slack.Token
-
-	client := resty.
-		NewWithClient(&http.Client{
-			Timeout: slackHTTPTimeout,
-			Transport: &http.Transport{
-				MaxConnsPerHost:     slackMaxConns,
-				MaxIdleConnsPerHost: slackMaxConns,
-			},
-		}).
-		SetHeader("Content-Type", "application/json").
-		SetHeader("Accept", "application/json").
-		SetHeader("Authorization", token)
-
-	// APIURL parameter is set only in tests
-	if endpoint := c.Slack.APIURL; endpoint != "" {
-		client.SetHostURL(endpoint)
-	} else {
-		client.SetHostURL("https://slack.com/api/")
-		client.OnAfterResponse(onAfterResponseSlack)
-	}
-
-	return SlackBot{
-		client:      client,
-		clusterName: clusterName,
-		webProxyURL: webProxyURL,
-	}, nil
 }
 
 // onAfterResponseSlack resty error function for Slack
